+++ js/app/services/erpHttpInterceptor.js
@@ -1,19 +1,5 @@
 angular
     .module('Erp')
-    .factory('dataFactory', ['$http', dataFactory]);
-
-function dataFactory($http) {
-    return {
-        getSharedData: function() {
-            return $http.get(ERPApp.baseAPIPath, {
-                params: { _do: 'getSharedData' }
-            });
-        }
-    };
-}
-
-angular
-    .module('Erp')
     .factory('erpHttpInterceptor', ['$q', '$rootScope', '$injector',
         function ($q, $rootScope, $injector) {
             $rootScope.http = null;